✨ Background Timer 로직 개선

+++ BreakTimer/BreakTimer Watch App/Utils/TimeFormatter.swift
@@ -7,10 +7,10 @@
 
 import Foundation
 
-extension Int {
+extension Double {
     func convertToTimeFormat() -> String {
-        let minute = self / 60
-        let second = self % 60
+        let minute = Int(self) / 60
+        let second = Int(self) % 60
         
         return String(format: "%02d:%02d", minute, second)
     }
